+++ packages/studio/src/components/AppSideMenu/messages.tsx
@@ -1,6 +1,7 @@
 import { defineMessages } from 'react-intl';
 
 export default defineMessages({
+  assets: 'Assets',
   details: 'Details',
   editor: 'Editor',
   resources: 'Resources',