+++ packages/studio/src/components/Assets/components/AssetPreview/messages.tsx
@@ -1,0 +1,6 @@
+import { defineMessages } from 'react-intl';
+
+export default defineMessages({
+  download: 'Download',
+  notSupported: 'This asset can not be previewed.',
+});