+++ CHANGELOG.md
@@ -13,10 +13,14 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - App: Add minimal `robots.txt`.
 - CLI: Add support for block icons.
 - Server: Add support for block icons.
+- Studio: Add asset management page for apps. This can be used to preview, download, delete, and
+  update assets.
 
 ### Fixes
 
 - App: Fix app theme color.
+- Studio: Add asset management page for apps. This can be used to preview, download, delete, and
+  update assets.
 
 ## [0.12.5] - 2020-04-16